Addressed Trim warnings in core Sentry package
