+++ src/Sentry/Internal/AotHelper.cs
@@ -3,6 +3,7 @@ namespace Sentry.Internal;
 internal static class AotHelper
 {
     internal const string SuppressionJustification = "Non-trimmable code is avoided at runtime";
+    internal static bool IsTrimmed { get; }
 
     private class AotTester
     {
@@ -17,10 +18,18 @@ internal static class AotHelper
     static AotHelper()
     {
         var stackTrace = new StackTrace(false);
-        IsNativeAot = stackTrace.GetFrame(0)?.GetMethod() is null;
+        IsTrimmed = stackTrace.GetFrame(0)?.GetMethod() is null;
+        IsNativeAot = IsTrimmed;
     }
 #else
     // This is a compile-time const so that the irrelevant code is removed during compilation.
     internal const bool IsNativeAot = false;
+
+    [UnconditionalSuppressMessage("Trimming", "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code", Justification = AotHelper.SuppressionJustification)]
+    static AotHelper()
+    {
+        var stackTrace = new StackTrace(false);
+        IsTrimmed = stackTrace.GetFrame(0)?.GetMethod() is null;
+    }
 #endif
 }
